updating main_menu.py file to launcher.py
